# [perf]: More precise producer rate limiter